adjusting navigation
