+++ jest.config.js
@@ -13,6 +13,7 @@ module.exports = {
     "^@app/(.*)$": "<rootDir>/src/app/$1",
     "^@components/(.*)$": "<rootDir>/src/components/$1",
     "^@assets/(.*)$": "<rootDir>/src/assets/$1",
+    "^@atomic$": "<rootDir>/src/atomic",
     "^@atomic/(.*)$": "<rootDir>/src/atomic/$1",
     "^@utils$": "<rootDir>/src/utils",
     "^@utils/(.*)$": "<rootDir>/src/utils/$1",
