+++ package.json
@@ -22,6 +22,7 @@
     "postcss-advanced-variables": "5.0.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
+    "react-router-dom": "^7.1.1",
     "tailwind-variants": "0.3.0"
   },
   "devDependencies": {
