+++ src/app/modules/home/home.route.ts
@@ -1,0 +1,3 @@
+export const HomeRoutes = {
+  Base: "/",
+};
