+++ src/app/modules/home/index.ts
@@ -1,0 +1,2 @@
+export * from "./home.route";
+export * from "./home.page";
