+++ src/atomic/index.ts
@@ -1,3 +1,4 @@
 export * from "./obj.constants";
 export * from "./mol.button";
 export * from "./mol.card";
+export * from "./atm.typography";
